Use `debian:bullseye-slim` image as base to overcome problems
The `ghcr.io/itchysats/itchysats/taker` image uses `distroless` as base, which is just too barebones for this setup.
Initially we overcame issues by installing `chown` / `chmod` (needed for changing permissions for volume and added scripts) by copying them from `busybox`.
However, the health-check uses `dev/stdin` to determine if the check should be triggered and that is not available on `distroless` images and could not be easily overcome.