Custom electrs URL in case of Electrs problems
Allow the user to specify a different Electrs URL in case of problems
Without Electrs ItchySats cannot operate properly, which can mean that if a position was opened, but then Electrs has problems this position might get force-closed.
To allow users to operate even if the Bitcoind / Electrs setup on their embassy makes problems we allow specifying a custom Electrs URL and specify a public blockcstream instance as default.